feat: functional minimal implem
