+++ libs/backend-ddd/src/project/domain/spi.ts
@@ -12,5 +12,4 @@ export interface ProjectRepository {
 
 export interface ProjectSorterInterface {
   byPriority(projects: ProjectType[], codeNAF: string | undefined): ProjectType[]
-  bySector(projects: ProjectType[]): ProjectType[]
 }
