+++ libs/data/src/index.ts
@@ -5,7 +5,6 @@ export * from './program/types/program'
 export * from './type/publicodesTypes'
 export * from './project/types/shared'
 export * from './project/projectEligibility'
-export * from './project/projectSorter'
 export * from './operators/types/domain'
 export * from './operators/types/generatedShared'
 export * from './testimony/types/shared'
